Fix warning

+++ modules/native_streaming_server_module/src/native_streaming_server_impl.cpp
@@ -176,7 +176,7 @@ void NativeStreamingServerImpl::coreEventCallback(ComponentPtr& sender, CoreEven
 void NativeStreamingServerImpl::initWorkerPool()
 {
     workerPool = nullptr;
-    Int workerCount = 1;
+    size_t workerCount = 1;
 
     if (config.hasProperty("StreamingWorkerCount"))
     {
